requirements.txt
fastapi
uvicorn
SQLAlchemy
qdrant-client
apscheduler
python-dotenv
transformers
torch
Pillow
requests
numpy
